refactor: simulation/mod.rs wires, and stops implementing
A `mod.rs` should let a reader see the shape of the thing. This one was 1017
lines, and most of them were not wiring: 157 of them were a hand-written table
mapping experience points to a level, sitting between a component declaration
and the ECS schedule.

Four modules come out, and every one of them is code motion. Nothing is
rewritten and no behaviour changes.

    xp.rs      168  the experience curve, as a table, and the level it displays
    pose.rs    245  position, rotation, size, chunk, velocity, teleport target
    lookup.rs   77  the three entity lookups and the DeferredMap under them
    flight.rs   69  flight permission, flying speed, fall tracking, abilities()

`mod.rs` goes from 1017 to 559, of which 23 lines are the module documentation
it did not have before.

## Seams, not line counts

Each file is one answer to "what is this about", which is why they are uneven
sizes. `pose.rs` is the components almost every system reads and almost none
writes, and it knows nothing about packets or about the schedule. `lookup.rs`
is four things that all answer "find this entity again", including the deferred
map the other three are built on. `xp.rs` is a pure function of one `u16`.

The accumulation was visible in the ordering. `impl Display for Pitch` came
before `struct Pitch`; `impl Position` appeared three separate times, once
either side of `ChunkPosition` and `PLAYER_SPAWN_POSITION`, because each
addition landed wherever the cursor happened to be. Both are now in one place,
which is the only part of this change that reorders rather than moves.

## What is left in mod.rs, and why

Marker components with no behaviour, `add_entity`, and `SimModule`. That is
wiring, so it belongs here. `SimModule::module` is now three named steps rather
than 267 lines in a row:

    fn module(world: &World) {
        register_reflection(world);
        let prefabs = register_components(world);
        register_observers(world, prefabs);
    }

The order is load-bearing and unchanged: a relation has to be a registered
entity before an argument node can point at it, and an observer has to come
after the components it watches. `register_reflection` is specifically the
metadata flecs cannot derive on its own, which is glam's and valence's types
plus `EntitySize` as an opaque.

`prefabs` is now a return value and a parameter. It used to be a `let` binding
that four observers closed over further down one long body; passing it says
that dependency out loud, and `MetadataPrefabs` is `Copy` so `world.set` still
takes its own.

`simulation/mod.rs` also had no module documentation at all, which for the file
whose job is to say what the module is was the larger omission. It now opens
with what the simulation is, that this file is wiring and its registration
order is load-bearing, and one line per submodule saying what subject each
owns. The last line is the one that matters for anyone adding to this tree:
nothing here knows which game is being played, and that is what `events/` is
for.

## Public paths did not move

Every item is re-exported at `simulation::` where it was declared before.
Which file a component lives in is an implementation detail; the path a game
module imports it by is not.

I checked this rather than assuming it: I extracted all 38 `pub` items from
`mod.rs` as it is on main, generated a test that names each one at
`hyperion::simulation::<Item>`, and compiled it against this branch. It builds,
so no downstream import can have broken. The generated file was temporary and
is not part of this change.

## Two consequences worth naming

`handlers.rs` read `pose.position` and `query.yaw.yaw` directly. Those fields
are private, and it could only see them because it was a child of the module
declaring the types. It is now a sibling, so both go through the `Deref` these
types already derive, which is what the line two below the first one was
already doing with `pose.y`.

`net/intermediate.rs` is reformatted. That is my own debt from #1013: renaming
`hyperion_proto` to `hyperion_proxy_proto` lengthened those lines past the
wrap, and I ran test, lint, doc and unused-deps on that PR but not `fmt`. Fixed
here rather than left.

Moving the `Position` doc into `pose.rs` broke its two intra-doc links to
`Player` and `Npc`, which stayed behind in `mod.rs`. `nix run .#doc` reported
them as warnings rather than errors, so they are the kind of thing a green
build hides; they are now `super::`-qualified. Two stale prose references went
the same way: a comment saying `EntitySize` "registers below as an opaque" now
names `super::register_reflection`, which is where that registration actually
moved to.

## Verified

- `nix run .#test` rc=0, 629 tests, unchanged from main.
- `nix run .#lint` rc=0, clean with no new allows.
- `nix run .#doc` rc=0, and no unresolved intra-doc link in any of the five
  files this change touches.
- `nix run .#fmt -- --check`: the only diff it reports is the pre-existing
  `adapter.rs` one below. Every file here is clean.
- The generated path-stability check above, rc=0.

All four were run against the exact tree this commit contains, after the last
edit rather than before it.

## Not done here

`nix run .#fmt -- --check` is still red on `events/smash/src/adapter.rs`, which
is pre-existing on main and tracked as ENG-10843. `cargo fmt --all` wanted to
fix it as a side effect; I reverted that, because the file sits next to work
another agent currently holds.

## Follow-up filed

`simulation/inventory.rs` is the next oversized file at 1572 lines, and it is
not a file-splitting problem: roughly 950 of those lines are the single function
`handle_click_slot`. While looking at it I found what looks like the root cause
of the kit slot symptom and filed it as ENG-10847. `hyperion-inventory` spells
absolute slots and hotbar indices both as `u16` with one undistinguished
conversion between them, `set_cursor` takes a hotbar index while
`get_cursor_index` returns an absolute slot, and `36` is written down in three
places. Not fixed here: the fix changes `set_hotbar`'s signature, whose callers
are in `events/smash/` files another agent currently holds for exactly this bug.

Authored by Claude (Anthropic AI agent).
